+++ Silicon/Nvidia/NvidiaPkg/Drivers/SmBiosTableDxe/SmBiosTableDxe.c
@@ -187,7 +187,7 @@ ProcessorInfoUpdateSmbiosType4 ()
     UINT8 GhzValue[2];
 
     // Split GHz Freq
-    GhzValue[0] = CpuFreq / 1000;
+    GhzValue[0] = (UINT8)(CpuFreq / 1000);
     GhzValue[1] = (CpuFreq / 10) % 100;
 
     // Append GHz Freq
@@ -195,8 +195,8 @@ ProcessorInfoUpdateSmbiosType4 ()
   }
 
   // Update Max & Current Speed
-  mProcessorInfoType4.MaxSpeed     = CpuFreq;
-  mProcessorInfoType4.CurrentSpeed = CpuFreq;
+  mProcessorInfoType4.MaxSpeed     = (UINT16)CpuFreq;
+  mProcessorInfoType4.CurrentSpeed = (UINT16)CpuFreq;
 
   // Update Core Count
   mProcessorInfoType4.CoreCount        = FixedPcdGet32 (PcdCoreCount);
@@ -312,7 +312,7 @@ MemDevInfoUpdateSmbiosType17 (IN UINT64 SystemMemorySize)
   mMemDevInfoType17.MemoryType = FixedPcdGet32 (PcdSmbiosMemoryType);
 
   // Update Memory Size
-  mMemDevInfoType17.Size = SystemMemorySize / 0x100000;
+  mMemDevInfoType17.Size = (UINT16)(SystemMemorySize / 0x100000);
 
   // Register SmBios Structure
   LogSmbiosData ((EFI_SMBIOS_TABLE_HEADER *)&mMemDevInfoType17, mMemDevInfoType17Strings, NULL);
@@ -323,7 +323,7 @@ MemArrMapInfoUpdateSmbiosType19 (IN UINT64 SystemMemorySize)
 {
   // Update Memory Start & End Address
   mMemArrMapInfoType19.StartingAddress = FixedPcdGet64 (PcdSystemMemoryBase) / 1024;
-  mMemArrMapInfoType19.EndingAddress   = (SystemMemorySize + FixedPcdGet64 (PcdSystemMemoryBase) - 1) / 1024;
+  mMemArrMapInfoType19.EndingAddress   = (UINT32)((SystemMemorySize + FixedPcdGet64 (PcdSystemMemoryBase) - 1) / 1024);
 
   // Register SmBios Structure
   LogSmbiosData ((EFI_SMBIOS_TABLE_HEADER *)&mMemArrMapInfoType19, mMemArrMapInfoType19Strings, NULL);
